Frontend: Split code in Home component into several modules
