+++ vue-app/src/api/abi.ts
@@ -1,0 +1,5 @@
+import { abi as FundingRoundFactory } from '../../../contracts/build/contracts/FundingRoundFactory.json'
+import { abi as FundingRound } from '../../../contracts/build/contracts/FundingRound.json'
+import { abi as ERC20 } from '../../../contracts/build/contracts/ERC20Detailed.json'
+
+export { FundingRoundFactory, FundingRound, ERC20 }
